requirements.txt
aiohttp==3.4.4
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==18.2.0
bottle==0.12.16
certifi==2019.3.9
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
Django==2.2.2
future==0.17.1
idna==2.7
image==1.5.27
ipython-genutils==0.2.0
jsonschema==3.0.1
jupyter-core==4.5.0
kiwisolver==1.1.0
lxml==4.3.4
matplotlib==3.0.3
multidict==4.4.2
nbformat==4.4.0
numpy==1.16.3
pandas==0.24.2
pep8==1.7.1
Pillow==6.0.0
plotly==3.10.0
psycopg2-binary==2.8.3
pycparser==2.19
pycryptodome==3.8.2
pyparsing==2.4.0
pyrsistent==0.15.2
python-dateutil==2.8.0
python-telegram-bot==11.1.0
pytz==2019.1
requests==2.21.0
retrying==1.3.3
six==1.12.0
sqlparse==0.3.0
telepot==12.7
traitlets==4.3.2
urllib3==1.24
virtualenv==16.0.0
yarl==1.2.6
